Added categorie check

+++ src/main/java/me/TechsCode/TechDiscordBot/module/cmds/GetReleaseCommand.java
@@ -7,6 +7,7 @@ import me.TechsCode.TechDiscordBot.module.CommandModule;
 import me.TechsCode.TechDiscordBot.objects.DefinedQuery;
 import me.TechsCode.TechDiscordBot.objects.Query;
 import me.TechsCode.TechDiscordBot.util.TechEmbedBuilder;
+import net.dv8tion.jda.api.entities.Category;
 import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.Role;
 import net.dv8tion.jda.api.entities.TextChannel;
@@ -20,9 +21,14 @@ import java.util.stream.Collectors;
 
 public class GetReleaseCommand extends CommandModule {
 
-    private final DefinedQuery<Role> STAFF_ROLE = new DefinedQuery<Role>() {
+    private final DefinedQuery<Role> SUPPORT_ROLES = new DefinedQuery<Role>() {
         @Override
-        protected Query<Role> newQuery() { return bot.getRoles("Staff"); }
+        protected Query<Role> newQuery() { return bot.getRoles("Supporter", "Senior Supporter", "Assistant", "Developer", "\uD83D\uDCBB Coding Wizard"); }
+    };
+
+    private final DefinedQuery<Category> SUPPORT_CATEGORIES = new DefinedQuery<Category>() {
+        @Override
+        protected Query<Category> newQuery() { return bot.getCategories("tickets", "paid plugin support"); }
     };
 
     public GetReleaseCommand(TechDiscordBot bot) {
@@ -41,7 +47,7 @@ public class GetReleaseCommand extends CommandModule {
 
     @Override
     public CommandPrivilege[] getCommandPrivileges() {
-        return new CommandPrivilege[] { CommandPrivilege.enable(STAFF_ROLE.query().first()) };
+        return new CommandPrivilege[] { CommandPrivilege.enable(SUPPORT_ROLES.query().first()) };
     }
 
     @Override
@@ -66,22 +72,31 @@ public class GetReleaseCommand extends CommandModule {
     public void onCommand(TextChannel channel, Member m, SlashCommandEvent e) {
         String plugin = e.getOption("plugin").getAsString();
 
-        e.reply("Getting release... please wait.").queue(q -> {
-            GithubRelease release = GitHubUtil.getLatestRelease(plugin);
+        if (channel.getParent() == SUPPORT_CATEGORIES.query().first()) {
+            e.reply("Getting release... please wait.").queue(q -> {
+                GithubRelease release = GitHubUtil.getLatestRelease(plugin);
 
-            if(release == null) {
-                q.editOriginal("**Failed!** Could not get the release!\n\n**Possible reasons:**\n- The repo isn't valid.\n- There is no release in the repo.\n- Github died.").queue();
-            } else if (release.getFile() != null) {
-                q.editOriginal(release.getFile(), plugin + ".jar")
-                        .queue(msg2 -> release.getFile().delete());
-                q.editOriginalEmbeds(
-                        new TechEmbedBuilder(release.getRelease().getName())
-                                .text("```" + (release.getRelease().getBody().isEmpty() ? "No changes specified." : release.getRelease().getBody().replaceAll(" \\|\\| ", "\n")) + "```")
-                                .build()
-                ).queue();
-            } else {
-                q.editOriginal("**Failed!** Could not get the file!\n\n**Possible reasons:**\n- Eazy messed up.\n- The release has no files for some reason.\n- GitHub died.").queue();
-            }
-        });
+                if (release == null) {
+                    q.editOriginal("**Failed!** Could not get the release!\n\n**Possible reasons:**\n- The repo isn't valid.\n- There is no release in the repo.\n- Github died.").queue();
+                } else if (release.getFile() != null) {
+                    q.editOriginal(release.getFile(), plugin + ".jar")
+                            .queue(msg2 -> release.getFile().delete());
+                    q.editOriginalEmbeds(
+                            new TechEmbedBuilder(release.getRelease().getName())
+                                    .text("```" + (release.getRelease().getBody().isEmpty() ? "No changes specified." : release.getRelease().getBody().replaceAll(" \\|\\| ", "\n")) + "```")
+                                    .build()
+                    ).queue();
+                } else {
+                    q.editOriginal("**Failed!** Could not get the file!\n\n**Possible reasons:**\n- Eazy messed up.\n- The release has no files for some reason.\n- GitHub died.").queue();
+                }
+            });
+        }else{
+            e.replyEmbeds(
+                    new TechEmbedBuilder("Get Release - Error")
+                            .text("You can not use this command in this channel.\n\n**Available channels:**\n  - Support channels\n  - Tickets")
+                            .error()
+                            .build()
+            ).setEphemeral(true).queue();
+        }
     }
 }
